[Refactor] Refactor vector index codes
