+++ be/src/storage/index/vector/vector_index_builder.cpp
@@ -14,9 +14,6 @@
 
 #include "storage/index/vector/vector_index_builder.h"
 
-#include <memory>
-
-#include "column/array_column.h"
 #include "fs/fs_util.h"
 #include "storage/index/index_descriptor.h"
 
@@ -24,7 +21,7 @@ namespace starrocks {
 
 Status VectorIndexBuilder::flush_empty(const std::string& segment_index_path) {
     ASSIGN_OR_RETURN(auto empty_file, fs::new_writable_file(segment_index_path));
-    RETURN_IF_ERROR(empty_file->append(MARK_WORD));
+    RETURN_IF_ERROR(empty_file->append(IndexDescriptor::mark_word));
     RETURN_IF_ERROR(empty_file->flush(WritableFile::FLUSH_SYNC));
     return empty_file->close();
 }
